+++ utils/start_dashboard.py
@@ -1,4 +1,4 @@
-from tulip_dashboard.server import start_server
+from airo_tulip_dashboard.server import start_server
 
 if __name__ == '__main__':
     start_server()